test-requirements.txt
pylint~=2.2
pytest~=4.1
pytest-aiohttp~=0.3.0
